requirements.txt
Flask
gunicorn
google-cloud-storage